requirements.txt
pyTelegramBotAPI==4.1.1
python-decouple==3.5
bs4==0.0.1
loguru==0.6.0
python-telegram-bot-calendar==1.0.5